// frontend/src/shared/api/openapi/models/ProductResponseV2Dto.ts
export interface ProductResponseV2Dto {
    /**
     * 
     * @type {number}
     * @memberof ProductResponseV2Dto
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof ProductResponseV2Dto
     */
    article: string;
    /**
     * 
     * @type {boolean}
     * @memberof ProductResponseV2Dto
     */
    isCustom: boolean;
    /**
     * 
     * @type {string}
     * @memberof ProductResponseV2Dto
     */
    logoUrl: string;
    /**
     * 
     * @type {string}
     * @memberof ProductResponseV2Dto
     */
    gender: ProductResponseV2DtoGenderEnum;
    /**
     * 
     * @type {number}
     * @memberof ProductResponseV2Dto
     */
    favoriteCount: number;
    /**
     * 
     * @type {number}
     * @memberof ProductResponseV2Dto
     */
    soldCount: number;
    /**
     * Популярность 0–100
     * @type {number}
     * @memberof ProductResponseV2Dto
     */
    popularityScore: number;
    /**
     * 
     * @type {string}
     * @memberof ProductResponseV2Dto
     */
    currency: ProductResponseV2DtoCurrencyEnum;
    /**
     * 
     * @type {number}
     * @memberof ProductResponseV2Dto
     */
    minPrice: number;
    /**
     * 
     * @type {number}
     * @memberof ProductResponseV2Dto
     */
    maxPrice: number;
    /**
     * 
     * @type {Date}
     * @memberof ProductResponseV2Dto
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof ProductResponseV2Dto
     */
    updatedAt: Date;
    /**
     * 
     * @type {number}
     * @memberof ProductResponseV2Dto
     */
    brandId: number;
    /**
     * 
     * @type {number}
     * @memberof ProductResponseV2Dto
     */
    categoryId: number;
    /**
     * 
     * @type {number}
     * @memberof ProductResponseV2Dto
     */
    modelId: number;
    /**
     * Source
     * @type {string}
     * @memberof ProductResponseV2Dto
     */
    mainSource: ProductResponseV2DtoMainSourceEnum;
    /**
     * 
     * @type {Array<Source>}
     * @memberof ProductResponseV2Dto
     */
    sources: Array<Source>;
    /**
     * 
     * @type {string}
     * @memberof ProductResponseV2Dto
     */
    title?: string;
}


/**
 * @export
 */
export const ProductResponseV2DtoGenderEnum = {
    Male: 'MALE',
    Female: 'FEMALE',
    Unisex: 'UNISEX',
    Child: 'CHILD',
    Baby: 'BABY',
    MiddleChild: 'MIDDLE_CHILD',
    EldestChild: 'ELDEST_CHILD'
} as const;
export type ProductResponseV2DtoGenderEnum = typeof ProductResponseV2DtoGenderEnum[keyof typeof ProductResponseV2DtoGenderEnum];

/**
 * @export
 */
export const ProductResponseV2DtoCurrencyEnum = {
    Rub: 'RUB',
    Cny: 'CNY',
    Eur: 'EUR',
    Usd: 'USD'
} as const;
export type ProductResponseV2DtoCurrencyEnum = typeof ProductResponseV2DtoCurrencyEnum[keyof typeof ProductResponseV2DtoCurrencyEnum];

/**
 * @export
 */
export const ProductResponseV2DtoMainSourceEnum = {
    Dewu: 'DEWU',
    Poizon: 'POIZON',
    Goat: 'GOAT',
    Shihuo: 'SHIHUO'
} as const;
export type ProductResponseV2DtoMainSourceEnum = typeof ProductResponseV2DtoMainSourceEnum[keyof typeof ProductResponseV2DtoMainSourceEnum];